Delete attribute2

+++ src/uml/Driver.java
@@ -79,29 +79,41 @@ public class Driver {
                             String attributeName = scan.next();
                             Attribute attribute = new Attribute(attributeName);
                             classToAddAtt.addAttribute(attribute);
+                            System.out.print("You added attribute \"" + attributeName + "\" to class \"" + classToAddAtt.getName() + "\"\n"); 
                         }
                         break;
                     case "delete attribute":
-                        // Call find attribute method
-                        Class classToDelAtt = findClass();
-                        // Prompt user to delete an attribute
-                        if (classToDelAtt != null) {
-                            System.out.print("Enter attribute name to delete: ");
-                            String attToDel = scan.next();
-                            System.out.print("Are you sure you want to delete \"" + attToDel + "\"? (y/n) ");
-                            String answer = scan.next();
-                            // If the user wants to delete an attribute, proceed to do so
-                            if (answer.equals("y")) {
-                                Attribute deletedAtt = classToDelAtt.findAtt(attToDel);
-                                classToDelAtt.deleteAttribute(deletedAtt);
-                                System.out.print("Attribute \"" + attToDel + "\" has been deleted");
-                            }
-                        }
+                         // Call find class method to find the class the attribute is
+                         Class classToDelAtt = findClass();
+                         Boolean attBool = false;
+                         // Prompt user to delete an attribute
+                         if (classToDelAtt != null) {
+                             // while bool is false continue loop until the user does not want to delete an attribute
+                             // or until an attribute is deleted 
+                             // (this may be changed so the user can type exit to exit the program as well)
+                             while(attBool.equals(false)){
+                                 System.out.print("Enter attribute name to delete: ");
+                                 String attToDel = scan.next();
+                                 Attribute deletedAtt = classToDelAtt.findAtt(attToDel);
+                                 if(deletedAtt != null){
+                                     System.out.print("Are you sure you want to delete \"" + attToDel +"\"? (y/n) ");
+                                     String answer = scan.next();
+                                     // If the user wants to delete an attribute, proceed to do so
+                                     if (answer.equals("y")) {
+                                         classToDelAtt.deleteAttribute(deletedAtt);
+                                         System.out.print("Attribute \"" + attToDel + "\" has been deleted \n");
+                                         attBool = true;
+                                     }
+                                     // if user types n, break out of loop and bring back prompt for new command
+                                     else{
+                                         break;
+                                     }
+                                 }
+                             }
+                         }
                         break;
-
                     case "rename attribute":
-
-                        // Show user classes and attributes before asking for input
+                         // Show user classes and attributes before asking for input
                         listClasses();
                         System.out.print("Enter the class that contains the attribute: ");
                         String classWithAttName = scan.next();
@@ -126,7 +138,6 @@ public class Driver {
                             }
                         }
                         break;
-
                     case "add relationship":
                         System.out.println("Enter source class name: ");
                         String sourceName = scan.next();
@@ -242,6 +253,10 @@ public class Driver {
         }
     }
 
+    /**
+     *
+     * List all classes and their accompanying attributes
+     */
     private static Relationship findRelationship() {
         System.out.print("Enter Relationship ID: ");
         String relationToFind = scan.next();
